refactor(ui): extract composer into chrome/thread/composer.tsx
Fourth extraction (Task 1): useComposer owns all composer state and the
send/queue/prefill/autogrow mechanics (still called from VendoThread so
draft and queued survive the landing/transcript flip); Composer is the
matching form presentation. Behavior-preserving.
